Create Decimal to Binary Conversion
Write a C++ function that performs decimal to binary conversion. The algorithm for this conversion is to repeatedly divide the decimal number by 2, until it is 0. Each division produces a remainder of 0 or 1, which becomes a digit in the binary number. For example, if we want the binary representation of decimal 13, we would find it with the following

series of divisions:

13/2 = 6           remainder 1

6/2 = 3             remainder 0

3/2 = 1             remainder 1

1/2 = 0             remainder 1

 

Thus, the binary representation of 13 (decimal) is 1101. The only problem with this algorithm is that the first division generates the low-order binary digit, the next division generates the second-order digit, and so on, until the last division produces the high order digit. Thus, if we output the digits as they are generated, they will be in reverse order. You should use recursion to reverse the order of output.