Create  inLanguage
Write a C++ function to recognize strings in this language, called inLanguage, that uses both a queue and a stack. The function takes a string as an argument and returns true if the input string is in the language, and false otherwise. Thus, as you traverse the input string, insert each character of S into a queue and each character of S’ into a stack.

If the input string is empty, the inLanguage function should print out the message "Empty string is not in the Language.", then return false. While if the input string is not an odd number of characters, it prints "A string with an odd number of characters is not in the Language.", then return false.

The signature of the function is defined as:

bool inLanguage(string & inputString);